+++ index.js
@@ -65,7 +65,7 @@ function fresh (reqHeaders, resHeaders) {
   if (noneMatch) {
     noneMatch = noneMatch.split(TOKEN_LIST_REGEXP)
     etagMatches = noneMatch.some(function (match) {
-      return match === '*' || match === etag || match === 'W/' + etag
+      return match === '*' || match === etag || match === 'W/' + etag || 'W/' + match === etag
     })
   }